split titan
